requirements.txt
# Core ML/DL frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0

# Computer vision & image processing
timm>=0.9.12
pillow>=10.1.0
opencv-python>=4.8.1

# Scientific computing
numpy>=1.24.3
scipy>=1.11.4
scikit-learn>=1.3.2

# Data validation & schemas
pydantic>=2.5.0

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0

# Optional: Model optimization
bitsandbytes>=0.41.3  # For 8-bit quantization if needed
sentencepiece>=0.1.99  # For tokenization

# Development
pytest>=7.4.3
black>=23.12.1
